chore: add create github release and send email workflows
- This adds create github release and send email workflows to the
project. The create github release workflow will create a github release
whenever a tag for a non-prerelease version is pushed. The body of the
release is composed of the last entry added to the changelog of the package
indicated by the tag. The send email workflow will send an email containing
the release notes of the release to the internal distribution list whenever a
release is published. Note that as lerna might apply several tags depending
on the released packages, it will generate as many emails as packages
that were released when there is a merge to `main` branch.