feat(react-native-analytics): update firebase mappings
- This updates Firebase integration mappings to be on par with the
web's GA4 integration.

BREAKING CHANGE: Now all integrations will use the same events API
that is being used by web applications. This means it is now possible
to use the bag and wishlist redux middlewares as they will
now be compatible with this implementation.
Also, AnalyticsService integration was removed as it
will not be supported in the future.