fix(analytics): fix bugs with `basic` mode in FirebaseAnalytics
This fixes a bug that happens in FirebaseAnalytics integration
when its `mode` option is set to `Basic` and a call to
`analytics.setConsent` is done later, the integration's own
`setConsent` method is not being called which prevents the
sending of the google consent mode values to firebase. Also, this fixes
the check for the call to firebaseAnalytics().setAnalyticsCollectionEnabled
to assume `Basic` as the default value for `mode` if none exists.